fix(recording): void cached permit decisions when the world moves
Apply the grok review findings on the R11 delta/timeline commit:

- A permit decision cached across a Pending retry no longer outlives
  its world: a watermark advance or an intervening E5 barrier (tracked
  by a barrier epoch) releases the cached lease and re-decides at the
  current watermark, so a barrier or revision advance can never commit
  a stale reuse (guarantees.md 5.3).
- A failed open resets the per-artifact delta/token/timeline state and
  clears CloseRequested: a stale delta base threw on the next artifact
  and a standing close request closed it immediately.
- Timeline appends stop at an evidence floor (MaxArtifactBytes minus
  MaxBlobBytes): diagnostics can never be what pushes the non-droppable
  lane into Incomplete(SizeLimit).
- The reader's delta consistency check compares in long arithmetic.
- recording-replay.md 4 / ADR 0015 now state the chain bound lives with
  the recording sink bounds, matching ADR 0016 and the implementation.
- New regression tests: failed-open state leak, Pending-permit world
  movement, independent byte-layout derivations for the delta and gap
  records.

+++ src/v2/Codec.Recording/ArtifactReader.cs
@@ -314,8 +314,8 @@ namespace SignalRouter.V2.Codec.Recording
                                 break;
                             }
 
-                            if (prefix + insert + suffix != resultLength ||
-                                prefix + suffix > baseBytes.Length)
+                            if ((long)prefix + insert + suffix != resultLength ||
+                                (long)prefix + suffix > baseBytes.Length)
                             {
                                 Degrade("Record " + recordCount + " carries an inconsistent delta.");
                                 break;
